monarch_messages

Module client

Structs

ClientActor

The generated behavior struct.

Enums

ClientMessage

Client messages. These define the messages that the controller can send to the client. The actual handling of these messages will be defined on the python side in the client implementation.

Exception

An exception to commanded execution, exposed to the client.

LogLevel

Log levels for ClientMessage::Log.

Traits

ClientMessageClient

The custom client trait for this message type.

ClientMessageHandler

The custom handler trait for this message type.