monarch_messages

Module controller

Structs

ControllerActor

The generated behavior struct.

DeviceFailure

Device operation failures.

Seq

The sequence number of the operation (message sent to a set of workers). Sequence numbers are generated by the client, and are strictly increasing.

WorkerError

Worker operation errors.

Enums

ControllerMessage

Controller messages. These define the contract that the controller has with the client and workers.

Ranks

Used to represent a slice of ranks. This is used to send messages to a subset of workers.

Traits

ControllerMessageClient

The custom client trait for this message type.

ControllerMessageHandler

The custom handler trait for this message type.