monarch_messages

Module worker

Structs

ActorCallParams

ActorMethodParams

ArgsKwargs

CallFunctionParams

Cloudpickle

Identifies a CallFunction target. Can either be a torch op or a Python global reference.

Factory

FunctionPath

Identifies a CallFunction target. Can either be a torch op or a Python global reference.

Ref

SeqError

An error associated with a seq number that failed to execute. Any defined value that has an error value will have an assocated SeqError that is the root cause of why that value has an error. A value may have this error because it was directly defined by the action associated with the sequence number, or if it was defined by another action that dependend on the failing one.

StreamRef

WorkerActor

The generated behavior struct.

WorkerParams

The parameters to spawn a worker actor.

Enums

CallFunctionError

When a worker runs any function, it may not succeed either because the function itself failed (Error) or because an input to the function already had an error value DependentError.

Reduction

Type of reduction for WorkerMessage::Reduce.

ResolvableFunction

StreamCreationMode

Controls what CUDA stream an actor will use.

WorkerMessage

Worker messages. These define the observable behavior of the worker, so the documentations here

Traits

WorkerMessageClient

The custom client trait for this message type.

WorkerMessageHandler

The custom handler trait for this message type.